-Added MySQL support (uncomment MySQL settings in application.ini) -Updated load.database.php to support MySQL -Bumped up version number

+++ application/scripts/load.database.php
@@ -61,20 +61,33 @@ if ('testing' != APPLICATION_ENV) {
     }
 }
 
-// Check to see if we have a database file already
-$options = $bootstrap->getOption('resources');
-$dbFile  = $options['db']['params']['dbname'];
-if (file_exists($dbFile)) {
-    unlink($dbFile);
+$databaseObjType = get_class($dbAdapter);
+if ($databaseObjType == 'Zend_Db_Adapter_Pdo_Mysql') {
+    $databaseType = 'mysql';
+} elseif ($databaseObjType == 'Zend_Db_Adapter_Pdo_Sqlite') {
+    $databaseType = 'sqlite';
+} else {
+    throw new Exception('Unsupported database type');
+}
+
+if ($databaseType == 'sqlite') {
+    // Check to see if we have a database file already
+    $options = $bootstrap->getOption('resources');
+    $dbFile  = $options['db']['params']['dbname'];
+    if (file_exists($dbFile)) {
+        unlink($dbFile);
+    }
 }
 
 // this block executes the actual statements that were loaded from
 // the schema file.
 try {
-    $schemaSql = file_get_contents(dirname(__FILE__) . '/schema.sqlite.sql');
+    $schemaSql = file_get_contents(dirname(__FILE__) . '/schema.'.$databaseType.'.sql');
     // use the connection directly to load sql in batches
     $dbAdapter->getConnection()->exec($schemaSql);
-    chmod($dbFile, 0666);
+    if ($databaseType == 'sqlite') {
+        chmod($dbFile, 0666);
+    }
 
     if ('testing' != APPLICATION_ENV) {
         echo PHP_EOL;
@@ -83,7 +96,7 @@ try {
     }
 
     if ($withData) {
-        $dataSql = file_get_contents(dirname(__FILE__) . '/data.sqlite.sql');
+        $dataSql = file_get_contents(dirname(__FILE__) . '/data.'.$databaseType.'.sql');
         // use the connection directly to load sql in batches
         $dbAdapter->getConnection()->exec($dataSql);
         if ('testing' != APPLICATION_ENV) {
